% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_framework.R
\name{eval_stats}
\alias{eval_stats}
\title{Calculate regression evaluation metrics}
\usage{
eval_stats(trues, preds, y_train, cf = 1.5, thr = 0.9, beta = 1)
}
\arguments{
\item{trues}{a vector of true values}

\item{preds}{a vector of predicted values}

\item{y_train}{a vector of training values}

\item{cf}{phi.control coef}

\item{thr}{relevance threshold}

\item{beta}{beta in F-measure}
}
\value{
a named vector of metrics RMSE, relevance-aware 
RMSE, and utility-based precision, recall and F-measure
}
\description{
Calculate regression metrics for imbalanced domains.
}
\author{
Nuno Moniz
}
