% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_weight.R
\name{get_phi}
\alias{get_phi}
\title{Calculate utility-based relevance}
\usage{
get_phi(y, phi.control)
}
\arguments{
\item{y}{vector of values to calculate relevance of}

\item{phi.control}{list of parameters as returned
by function \code{UBL::phi.control}}
}
\description{
Calculate relevance of values given a parametrization
of the relevance function. 
Most relevant: phi -> 1; less relevant: phi -> 0.
}
\seealso{
\code{\link[UBL]{phi}}, \code{\link[UBL]{phi.control}}
}
