% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_weight.R
\name{get_time_wts}
\alias{get_time_wts}
\title{Calculate temporally-biased re-sampling weights}
\usage{
get_time_wts(times, phi, rel.thr)
}
\arguments{
\item{times}{a vector of time-stamps}

\item{phi}{a vector of the relevance values of 
\code{df}'s target variable}

\item{rel.thr}{a relevance threshold above which an 
observation is considered relevant}
}
\value{
A vector of temporally-biased re-sampling weights, scaled
to fit within range [0,1].
}
\description{
Calculate weights for re-sampling with a temporal bias.
Most recent observations have weights that tend to 1,
while the oldest observations have weights that tend to 0
(meaning they are less likely to be kept).
Most recent observations: w -> 1; oldest: w -> 0.
}
\author{
Mariana Oliveira
}
