% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_framework.R
\name{int_util_evaluate}
\alias{int_util_evaluate}
\title{Evalute the results of an internal predictive workflow}
\usage{
int_util_evaluate(wfRes, eval.function = get("regressionMetrics",
  asNamespace("performanceEstimation")), y_train, ...)
}
\arguments{
\item{wfRes}{a data frame (or list of data frames) containing the results of
a predictive workflow with columns \code{trues} and \code{preds} containing
the real and predicted values, respectively}

\item{eval.function}{the function to be used to calculate error metrics from \code{wfRes}}

\item{y_train}{a vector of the whole training values}

\item{...}{parameters to pass to \code{eval.function}}
}
\value{
The results (or a list of results) of \code{eval.function} applied to 
the data frame (or list of data frames) in \code{wfRes}
}
\description{
Calculate evaluation metrics from the raw results of a workflow
}
