% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflows.R
\name{internal_workflow}
\alias{internal_workflow}
\title{A learning and prediction workflow with internal validation}
\usage{
internal_workflow(train, test, form, model = "lm", resample = NULL,
  resample.pars = NULL, internal.est = NULL,
  internal.est.pars = NULL, internal.evaluator = NULL,
  internal.eval.pars = NULL, metrics = NULL, metrics.max = NULL,
  stat = "MED", resample.grid = NULL, handleNAs = NULL,
  min_train = 2, nORp = 0.2, time = "time", site_id = "site",
  .full_intRes = FALSE, ...)
}
\arguments{
\item{train}{a data frame for training}

\item{test}{a data frame for testing}

\item{form}{a formula describing the model to learn}

\item{model}{the name of the algorithm to use}

\item{resample}{re-sampling technique to be used. Default is NULL.}

\item{resample.pars}{parameters to be passed to re-sample function.
Default is NULL.}

\item{internal.est}{character string identifying the internal estimator
function to use}

\item{internal.est.pars}{named list of internal estimator
parameters (e.g., tr.perc or nfolds)}

\item{internal.evaluator}{character string indicating internal 
evaluation function}

\item{internal.eval.pars}{named list of parameters to feed to internal 
evaluation function}

\item{metrics}{vector of names of two metrics to be used to determine the best 
parametrization (the second metric is only used in case of ties)}

\item{metrics.max}{vector of Booleans indicating whether each metric in 
parameter metrics should be maximized (TRUE) or minimized (FALsE)
for best results}

\item{stat}{parameter indicating summary statistic that should be 
used to determine the best internal evaluation metric: 
"MED" (for median) or "MEAN" (for mean)}

\item{resample.grid}{a data.frame with columns indicating
resample.pars to test using internal.est}

\item{handleNAs}{string indicating how to deal with NAs.
If "centralImput", training observations with at least 80\%
of non-NA columns, will have their NAs substituted by the mean
value and testing observatiosn will have their NAs filled in with
mean value regardless. Default is NULL.}

\item{min_train}{a minimum number of observations that must be
left to train a model. If there are not enough observations, 
predictions will be \code{NA}. Default is 2.}

\item{nORp}{a maximum number or fraction of columns/rows with missing
values above which a row/column will be removed from train before 
learning the model. Only works if \code{handleNAs} was
set to centralImputation. Default is 0.2.}

\item{time}{the name of the column in \code{train} and
\code{test} containing time-stamps}

\item{site_id}{the name of the column in \code{train} and
\code{test} containing location IDs}

\item{.full_intRes}{a Boolean indicating whether the full results
object for internal validation should be returned as well. Defaults to FALSE}

\item{...}{other parameters to feed to \code{model}}
}
\value{
a data frame containing time-stamps, location IDs,
true values and predicted values
}
\description{
A learning and prediction workflow that may deal
with NAs and use internal validation to parametrize a
re-sampling technique to balance an imbalanced regression problem.
}
