% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UBL_sampling_methods.R
\name{RandOverRegress}
\alias{RandOverRegress}
\title{Random over-sampling for imbalanced regression problems}
\usage{
RandOverRegress(form, dat, rel = "auto", thr.rel = 0.5,
  C.perc = "balance", repl = TRUE)
}
\arguments{
\item{form}{
    A formula describing the prediction problem.
  }

\item{dat}{
    A data frame containing the original imbalanced data set.
  }

\item{rel}{
    The relevance function which can be automatically ("auto") determined (the default) or may be provided by the user through a matrix with the interpolating points.
  }

\item{thr.rel}{
    A number indicating the relevance threshold above which a case is considered as belonging to the rare "class".
  }

\item{C.perc}{A vector containing the over-sampling percentage/s to apply to all/each 
"class" (bump) obtained with the relevance threshold. Replicas of the examples are are
 randomly added in each "class". If only one percentage is provided this value is reused 
 in all the "classes" that have values above the relevance threshold. A different percentage 
 can be provided to each "class". In this case, the percentages should be provided in 
 ascending order of target variable value. The over-sampling percentage(s), should be 
 numbers above 0, meaning that the important cases (cases above the threshold) are over-sampled 
 by the corresponding percentage. If the number 1 is provided then the number of extreme examples 
 will be doubled. 
 Alternatively, C.perc parameter may be set to "balance" or "extreme", cases where the 
 over-sampling percentages are automatically estimated to either balance or invert the
 frequencies of the examples in the "classes" (bumps).}

\item{repl}{
  A boolean value controlling the possibility of having repetition of examples when choosing the examples to repeat in the over-sampled data set. 
Defaults to TRUE because this is a necessary condition if the selected percentage is greater than 2. This parameter is only important when the over-sampling percentage is between 1 and 2. In this case, it controls if all the new examples selected from a given "class" can be repeated or not.
  }
}
\value{
The function returns a data frame with
  the new data set resulting from the application of the random over-sampling strategy.
}
\description{
This function is identical to the function of the same name available on the 
R package \code{UBL}. The only difference is in the requirements
imposed on the argument C.perc. \cr
This function performs a random over-sampling strategy for imbalanced 
regression problems. Basically a percentage of cases of the "class(es)" 
(bumps above a relevance threshold defined) selected by the user are randomly 
over-sampled. Alternatively, it can either balance all the existing "classes" 
(the default) or it can "smoothly invert" the frequency of the examples in each class.
}
\details{
The only difference between this function and the original function is in the requirements 
imposed on the argument C.perc. \cr This function performs a random over-sampling strategy for dealing 
with imbalanced regression problems. The new examples included in the new data set
 are randomly selected replicas of the examples already present in the original data set.
}
\references{
Paula Branco, Rita P. Ribeiro, Luis Torgo (2016)., 
UBL: an R Package for Utility-Based Learning, 
CoRR abs/1604.08079 [cs.MS], URL: http://arxiv.org/abs/1604.08079
}
\seealso{
\code{\link[UBL]{RandOverRegress}}, \code{\link{RandUnderRegress}}
}
