% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_methods.R
\name{randUnderRegress_ST}
\alias{randUnderRegress_ST}
\title{Biased under-sampling for imbalanced regression spatio-temporal problems}
\usage{
randUnderRegress_ST(form, dat, alpha = 0.5, beta = 0.9, rel = "auto",
  thr.rel = 0.5, epsilon = 1e-04, C.perc = "balance", repl = FALSE,
  type = "add", site_id = "site_id", time = "time",
  sites_sf = NULL, lon = NULL, lat = NULL, crs = NULL)
}
\arguments{
\item{form}{a model formula}

\item{dat}{the original training set (with the unbalanced distribution)}

\item{alpha}{weighting parameter for temporal and spatial
re-sampling probabilities. Default 0.5}

\item{beta}{weighting parameter for spatiotemporal weight and phi for
re-sampling probabilities. Default 0.9}

\item{rel}{relevance determined automatically (default) with uba package 
or provided by the user}

\item{thr.rel}{relevance threshold above which a case is considered as 
belonging to the rare "class"}

\item{epsilon}{minimum weight to be added to all observations. 
Default 1E-4}

\item{C.perc}{A vector containing the over-sampling percentage/s to apply to all/each 
"class" (bump) obtained with the relevance threshold. Replicas of the examples are are
 randomly added in each "class". If only one percentage is provided this value is reused 
 in all the "classes" that have values above the relevance threshold. A different percentage 
 can be provided to each "class". In this case, the percentages should be provided in 
 ascending order of target variable value. The over-sampling percentage(s), should be 
 numbers above 0, meaning that the important cases (cases above the threshold) are over-sampled 
 by the corresponding percentage. If the number 1 is provided then the number of extreme examples 
 will be doubled. 
 Alternatively, C.perc parameter may be set to "balance" or "extreme", cases where the 
 over-sampling percentages are automatically estimated to either balance or invert the
 frequencies of the examples in the "classes" (bumps).}

\item{repl}{allowed to perform sampling with replacement}

\item{type}{character string indicating the type of bias used. Default is "add".
More types to be added in future work}

\item{site_id}{the name of the column containing location IDs}

\item{time}{the column name of the time-stamp}

\item{sites_sf}{An sf obejct containing station and IDs and 
geometry points of the locations. As an alternative, provide
\code{lon}, \code{lat}, and \code{crs}}

\item{lon}{the name of the column containing the location's longitude}

\item{lat}{the name of the column containing the location's latitude}

\item{crs}{the code for the Coordinate Reference System}
}
\value{
The function returns a data frame with the new data set resulting 
from the application of the spatio-temporally biased under-sampling strategy.
}
\description{
Based on randUnderRegress (\code{R} package \code{UBL}). 
The function performs a random under-sampling strategy for
imbalanced regression problems with a bias based on spatio-temporal
contextual information. Essentially, a percentage of 
cases of the "class(es)" (bumps below a relevance threshold defined) 
selected by the user are randomly removed with a sampling bias
based on a spatio-temporal weight. Alternatively, the strategy 
can be applied to either balance all the existing "classes"" or to 
"smoothly invert" the frequency of the examples in each "class".
}
\references{
Paula Branco, Rita P. Ribeiro, Luis Torgo (2016)., 
UBL: an R Package for Utility-Based Learning, 
CoRR abs/1604.08079 [cs.MS], URL: http://arxiv.org/abs/1604.08079
}
\seealso{
\code{\link[UBL]{RandUnderRegress}}, \code{\link{sample_wts}}
}
