% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflows.R
\name{summarize_metrics}
\alias{summarize_metrics}
\title{Summarize metrics}
\usage{
summarize_metrics(int.res, metrics)
}
\arguments{
\item{int.res}{A list with results obtained by running estimates}

\item{metrics}{a list of metrics that should be summarized}
}
\value{
a named vector with median, IQR, mean, standard-deviation,
and number of non-NA values of each metric
}
\description{
Used for internal validation
}
